move profile constants to common
